backend/requirements.txt
# Core dependencies
tensorflow>=2.8.0
numpy>=1.19.5
Pillow>=9.0.0

# API dependencies
fastapi>=0.75.0
uvicorn>=0.17.0
python-multipart>=0.0.5
pydantic>=1.9.0

# Image processing
opencv-python>=4.5.5
scikit-image>=0.19.2

# Utilities
tqdm>=4.63.0
matplotlib>=3.5.1
